Reduced memory footprint and made CSV output the default
This patch is especially for people who don't like running out of memory or
disk space. The database is now written to disk and cleared on each file, rather
than saving it all to the end.

Also, CSV outputs are now the default. It didn't make much sense to calculate
everything and then throw it away.

Signed-off-by: Brian Warner <brian@bdwarner.com>