This commit dedicated to people who don't like running out of memory
So, we had a little problem until now.  On trivial files Water worked fine, but
on bigger files the internal sqlite database would expand and expand until the
system killed it.

Previously for each file, when a line was matched it created a new row in the
sqlite database, which was unsustainable unless you have infinite RAM. This was
done because sqlite doesn't have a built-in "ON DUPLICATE KEY UPDATE" feature,
and then all the dupes were collated at the end using GROUP BY in the SELECT
statement.

It turns out there's a hacktastic way to simulate this behavior, and it keeps
the memory consumption down at reasonable levels.  There are likely performance
implications, but it's better than losing the analysis partway through.

Signed-off-by: Brian Warner <brian@bdwarner.com>