Major rework to use git for search the repo instead of doing it manually
* This still requires testing to make sure it is faster! *

This is a major rework of how lines in the snapshot are matched against the git
log. Until now it was done manually, meaning that the git log was read entirely
into memory and we looked for matches.

This patch tests another way of doing it, which is to use git log to do the
searching.  It means a dramatic simplification of things, but also needs to be
verified against some real world repos to be sure it's an improvement.

Conveniently, this patch also adds timing to the script.

Signed-off-by: Brian Warner <brian@bdwarner.com>