Move in-memory database logic into main for loop
Previously we had a single database that we just reused each time the for loop
ran, and then emptied out at the end. Now that we're going to give
multithreading a shot, we need to be more careful about keeping data separate.
It seems to make more sense to create/destroy one sqlite database per file,
since according to the sqlite docs, these are private per instance.
https://www.sqlite.org/inmemorydb.html

With this change, it also no longer makes much sense to have the option to dump
the database, either.  So that has been eliminted.

Signed-off-by: Brian Warner <brian@bdwarner.com>