Ignore certain files, improved multithreading
Previous versions of Water referenced the ability to ignore common image files,
because it doesn't make sense to read binary data line-by-line and compare it
against the git repo.  This would seriously bias the statistics.  Unfortunately,
it was just a stub.

This patch enables Water to ignore common image formats, PDF, EPS, and PS
files, and LICENSE files.  It can be disabled, but that isn't smart.

Also, this patch adds support for disabling multithreading.

Finally, this patch disables multithreading anytime a verbosity option is used,
because the log messages are really confusing when they arrive out-of-order.

Signed-off-by: Brian Warner <brian@bdwarner.com>